+++ react/src/components/comparison-panel.tsx
@@ -5,7 +5,8 @@ import React, { ReactNode, useContext, useEffect, useRef } from 'react'
 
 import { sanitize } from '../navigation/links'
 import { Navigator } from '../navigation/navigator'
-import { HueColors, useColors } from '../page_template/colors'
+import { HueColors } from '../page_template/color-themes'
+import { useColors } from '../page_template/colors'
 import { row_expanded_key, useSetting, useSettings } from '../page_template/settings'
 import { groupYearKeys, StatGroupSettings } from '../page_template/statistic-settings'
 import { PageTemplate } from '../page_template/template'
