+++ react/src/components/related-button.tsx
@@ -4,7 +4,8 @@ import './related.css'
 import type_ordering_idx from '../data/type_ordering_idx'
 import type_to_type_category from '../data/type_to_type_category'
 import { Navigator } from '../navigation/navigator'
-import { HueColors, useColors } from '../page_template/colors'
+import { HueColors } from '../page_template/color-themes'
+import { useColors } from '../page_template/colors'
 import { relationship_key, useSetting } from '../page_template/settings'
 import { useUniverse } from '../universe'
 import { mixWithBackground } from '../utils/color'
