+++ react/src/components/sidebar.tsx
@@ -4,7 +4,8 @@ import '../style.css'
 import './sidebar.css'
 
 import { Navigator } from '../navigation/navigator'
-import { Theme, useColors, useCurrentTheme } from '../page_template/colors'
+import { Theme } from '../page_template/color-themes'
+import { useColors, useCurrentTheme } from '../page_template/colors'
 import { checkbox_category_name, SettingsDictionary, source_enabled_key, TemperatureUnit, useSetting, useSettingInfo, useStagedSettingKeys } from '../page_template/settings'
 import { useDataSourceCheckboxes } from '../page_template/statistic-settings'
 import { useMobileLayout } from '../utils/responsive'
