+++ react/src/components/table.tsx
@@ -5,7 +5,8 @@ import { load_ordering } from '../load_json'
 import './table.css'
 import { statisticDescriptor } from '../navigation/links'
 import { Navigator } from '../navigation/navigator'
-import { Colors, useColors } from '../page_template/colors'
+import { Colors } from '../page_template/color-themes'
+import { useColors } from '../page_template/colors'
 import { MobileArticlePointers, row_expanded_key, Settings, useSetting } from '../page_template/settings'
 import { useUniverse } from '../universe'
 import { is_historical_cd } from '../utils/is_historical'
